Ensure source is set via Syntax#register_mapping
Made Paraphrase.mappings a editable externally accessible so
registering a mapping doesn't require going through Paraphrase.register.

Add Paraphrase.[]= as a shortcut for setting mappings.

Restructure Query.paraphrases to accept a class instead of a symbol or
string to retrieve the class.  In ActiveRecord, this will be autoloaded.
Elsewhere, it should be required/loaded before the query class.