Add .paraphraser class_attribute to AR::Base
Instead of fetching the query class from a central store, store it on
the model class itself. This solves problems related to AR classes
reloading in development where the query class's source would reference
the previous model class.

If creating a subclass of Paraphrase::Query and the source is set by
:paraphrases, don't worry about setting :paraphraser on the
corresponding model.  In this use case, the class should be used
directly.