Support `ActionController::Parameters`
Remove filtering keys not in `keys` argument from `unfiltered_params`
using `#slice`. Since `keys` is being iterated over, only those values
will be in the resulting hash.

Handle indifferent access for the hash-like object passed to
`ParamsFilter` by defining `ParamsFilter#params` to return itself and
define its `#[]` to check `@params` for the symbol or string versions of
the key. `ActionController::Parameters#with_indifferent_access` is
deprecated in Rails 5 and removed in Rails 5.1. It can be converted to
an `ActiveSupport::HashWithIndifferentAccess` using `#to_h` after
calling `permit` to specify what keys to keep.  If `permit` is not
called, an empty `ActiveSupport::HashWithIndifferentAccess` is returned.

Misc:
* Change `Paraphrase::ParamsFilter#result` to a method; stop memoizing it.
* When defining keys through `Query.map`, define the default method on
  the `ParamsFilter` subclass of the `Query` subclass removing the need
  for a `respond_to?` check when filtering params.